Fix import error due to #71
The utils crate is re-exported from the main crate for use in debug
builds.

Debug builds are compiled by the macro to use certain functions from the
utils crate. This is different than in release mode, where the macro
itself uses the functions and does not require them at runtime. When the
debug code tries to import the utils crate, it fails because it is not
an explicit dependency in Cargo.toml.

Fixes #74

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>