Remove `syn` features that are not required
The `impl` proc macro depends on `syn`. Not all of the features enabled
in `syn`'s set of default features are required to build the `impl` proc
macro.

Disable default features in the `syn` dependency and only enable the
minimum set of features required to get the APIs needed by the `impl`
crate. This has a nice side effect of slightly speeding up compile
times.