Fix naming collisions in macros
Previously the macro code wasn't using absolute paths in the code it generated, which could result in naming collisions from items the caller had in scope. This fixes such by making imports have absolute paths.

Currently the 'rust_embed' crate isn't absolutely scoped, as I'm unaware of a way to handle that reliably in proc macros (i.e. when the user renames the 'rust_embed' crate in their Cargo.toml). If that should be addressed in this PR I'm more than open to doing such, but even with this in itself it's still good progress compared to what was present before.

Closes #229.