Replace `expect` with a safer alternative that returns `None` instead
Given the fact that this function returns a `Result`, in case of a
failure, the function should either return an `Err` or silent fail.

When using `expect`, the program just panics, so none of the above
situations happen.

I have replaced `expect` with `and_then`. Now the code silent fails
in case time before UNIX epoch is not supported, returning `None` instead.

Signed-off-by: Costin-Robert Sin <sin.costinrobert@gmail.com>