perf: Do not build glob matchers repeatedly when include-exclude feature is enabled
I've noticed in my profiles that if include-exclude feature is enabled, the call to ::iter() function is dominated by path filtering.
This commit fixes that by explicitly passing in a premade Globset for a given object.

Note that we're using std::sync::OnceLock to lazily store away premade Globsets, which bumps MSRV to 1.70. An alternative would be to use once_cell as an explicit dependency.

For my use case, in debug builds this takes down the matching time from 650ms to 6ms. I know you shouldn't benchmark debug builds, but rust-embed::iter is on my app's initialization path and that's how I noticed it in the first place. I'd expect release timings to be somewhat similar